chore: refactor command code
