+++ scripts/test.sh
@@ -7,7 +7,7 @@ echo "=== TEST ===" &&
 cargo nextest run &&
 echo "=== FORGOTTEN TODOS ===" &&
 # Requires ripgrep
-if rg -i -s -g '*.rs' 'TODO:|todo:|FIXME|fixme|dbg!|DEBUG:|FIXTURE:' .; then
+if rg -i -s -g '*.rs' 'TODO:|todo:|FIXME|fixme|dbg!|DEBUG:|FIXTURE:|TODO\s|todo\s' .; then
     exit 1
 fi
 echo "=== SUCCESS ===" &&
